dhruvi/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
httpx==0.27.0
python-dotenv==1.0.1
pinecone-client==4.1.0
google-generativeai==0.5.0
pydantic==2.7.1
streamlit==1.46.1
requests==2.32.4
qrcode==7.4.2
python-Levenshtein==0.25.0
fuzzywuzzy==0.18.0
flask==3.0.3
flask-cors==4.0.1
openai==1.30.1
opencv-python==4.9.0.80
pyzbar==0.1.9
